[MODIFY] move ad code to recipe screen
